+++ OzzCodeGen/CodeEngines/BaseAppInfraCodeEngine.cs
@@ -1,4 +1,5 @@
-﻿using System.IO;
+﻿using OzzCodeGen.CodeEngines.CsModelClass;
+using System.IO;
 using System.Text.Json.Serialization;
 using System.Xml.Serialization;
 
@@ -30,6 +31,32 @@ namespace OzzCodeGen.CodeEngines
         [JsonIgnore]
         public bool HasDifferentFolderForContracts => !string.IsNullOrWhiteSpace(InfrastructureFolder)
                                                    && !string.Equals(TargetFolder, InfrastructureFolder, System.StringComparison.InvariantCulture);
+
+
+        /// <summary>
+        /// Model class code engine is used to get information about the model classes, such as validator class, which can be used in the repository templates.
+        /// </summary>
+        [XmlIgnore]
+        [JsonIgnore]
+        public CSharpModelClassCodeEngine ModelClassCodeEngine
+        {
+            get
+            {
+                if (_modelClassEngine == null && Project != null)
+                {
+                    _modelClassEngine = Project.GetCodeEngine(EngineTypes.CsModelClassCodeEngineId) as CSharpModelClassCodeEngine;
+                }
+                return _modelClassEngine;
+            }
+        }
+        private CSharpModelClassCodeEngine _modelClassEngine;
+
+        [XmlIgnore]
+        [JsonIgnore]
+        public string QueryParamNamespaceName
+        {
+            get { return ModelClassCodeEngine?.QueryParamNamespaceName ?? string.Empty; }
+        }
 
         /// <summary>
         /// Namespace for generated base classes or contract interfaces (for example <c>IViewModel</c>, <c>IAsyncCommand</c>, <c>INavigationService</c>).